+++ config/constants.py
@@ -1,0 +1,6 @@
+PASSENGER_GENERATION_INTERVAL = 1  # seconds
+BUS_MOVE_DELAY = 2  # seconds
+STOP_WAIT_TIME = 2  # seconds
+STOP_RADIUS = 25  # Radius for larger stops
+SMOOTH_MOVE_INTERVAL = 50  # milliseconds
+STEPS_PER_ROUTE = 20  # Number of steps between stops for smooth movement
