--- config/settings.py
@@ -1,7 +1,0 @@
-# # config/settings.py
-
-# # Configuration settings
-# BUS_CAPACITY = 10          # Maximum passengers per bus
-# NUM_STOPS = 8              # Number of stops on the route
-# INITIAL_PASSENGER_RATE = 0.5  # Rate of passenger generation
-# SIMULATION_SPEED = 1         # Speed multiplier for the simulation
