--- logic/utils.py
@@ -1,9 +1,0 @@
-# # logic/utils.py
-
-# from random import randint
-
-# def generate_random_passenger(start_stop, num_stops):
-#     end_stop = randint(0, num_stops - 1)
-#     while end_stop == start_stop:
-#         end_stop = randint(0, num_stops - 1)
-#     return end_stop
