+++ models/bus.py
@@ -28,17 +28,11 @@ class Bus:
         transit_passengers = [
             p for p in self.passengers if p.intermediate_stop == self.current_stop
         ]
-        # transit_passengers = []
-        # for passenger in self.passengers:
-        #     if(passenger.intermediate_stop):
-        #         transit_passengers = [
-        #             p for p in self.passengers if p.intermediate_stop == self.current_stop
-        #         ]
+
         # Remove these passengers from the bus
         self.passengers = [
             p for p in self.passengers if p.end != self.current_stop and p.intermediate_stop != self.current_stop
         ]
-        # print("Passengers on bus: ", str(self.passengers))
 
         for passenger in deboarded_passengers:
             passenger.status = "Deboarded"
