--- models/busStop.py
@@ -1,10 +1,0 @@
-# from models.passenger import Passenger
-
-# class BusStop:
-#     def __init__(self, stop_id):
-#         self.stop_id = stop_id
-#         self.waiting_passengers = []
-
-#     def add_passenger_request(self, end_stop):
-#         passenger = Passenger(self.stop_id, end_stop)
-#         self.waiting_passengers.append(passenger)
